main: move verbose manifest logging to project.py
This kind of verbose detail about the parsed manifest is not always
useful to others.

E.g. downstream in the nRF Connect SDK, this is cluttering up the
output of a west extension that compares mainline zephyr's manifest
repository revisions with NCS's. This seems like an annoyance to me,
so this logging shouldn't be global.

We do, however, need it for debugging when resolving the manifest
and running 'west -vv update'.
